Migrate dialog into OneStopDialog
