+++ client/package.json
@@ -32,7 +32,6 @@
     "npm": "^6.14.7",
     "prop-types": "^15.6.2",
     "react": "^16.7.0",
-    "react-a11y-dialog": "^4.1.1",
     "react-animate-height": "^2.0.7",
     "react-aria-live": "^2.0.4",
     "react-browser-detection": "^1.2.1",
