+++ client/src/components/cart/Cart.jsx
@@ -6,22 +6,22 @@ import Button from '../common/input/Button'
 import {boxShadow} from '../../style/defaultStyles'
 import {identifyProtocol} from '../../utils/resultUtils'
 import clearIcon from 'fa/ban.svg'
-import {times, SvgIcon} from '../common/SvgIcon'
+// import {times, SvgIcon} from '../common/SvgIcon'
 import {fontFamilySerif} from '../../utils/styleUtils'
 import ScriptDownloader from './ScriptDownloader'
 import {FEATURE_CART} from '../../utils/featureUtils'
 import CartListItem from './CartListItem'
 import {PAGE_SIZE} from '../../utils/queryUtils'
-// import {Confirmation, useConfirmation} from '../common/dialog/OneStopDialog'
+import {Confirmation} from '../common/dialog/OneStopDialog'
 import {
   useDisclosure,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
+  // Modal,
+  // ModalOverlay,
+  // ModalContent,
+  // ModalHeader,
+  // ModalFooter,
+  // ModalBody,
+  // ModalCloseButton,
 } from "@chakra-ui/core";
 
 const styleCenterContent = {
@@ -55,17 +55,17 @@ const styleCartActionsTitle = {
   padding: 0,
 }
 
-const styleButton = {
-  padding: '0.309em',
-  margin: '0.105em',
-  borderRadius: '0.309em',
-  fontSize: '1em',
-}
+// const styleButton = {
+//   padding: '0.309em',
+//   margin: '0.105em',
+//   borderRadius: '0.309em',
+//   fontSize: '1em',
+// }
 
-const styleButtonFocus = {
-  outline: '2px dashed black',
-  outlineOffset: '2px',
-}
+// const styleButtonFocus = {
+//   outline: '2px dashed black',
+//   outlineOffset: '2px',
+// }
 // export default class Cart extends React.Component {
 export default function Cart({
   featuresEnabled,
@@ -101,6 +101,7 @@ export default function Cart({
   //   },
   //   noText: 'Never mind',
   // })
+
 
   //only show granules for this page
   const allowed = Object.keys(selectedGranules).slice(
@@ -188,6 +189,7 @@ export default function Cart({
     cartListCustomActions.push(clearAllAction)
   }
 
+
 // <Modal isOpen={isOpen} onClose={onClose}>
 //   <ModalOverlay />
 //   <ModalContent>
@@ -228,39 +230,43 @@ export default function Cart({
 //   </ModalContent>
 //   </ModalOverlay>
 // </Modal>
+// <Modal isOpen={isOpen} onClose={onClose} isCentered >
+//   <ModalOverlay backgroundColor="#5d5d5d94">
+//   <ModalContent maxWidth="28em" borderRadius='0.309em' padding="0.309em">
+//     <ModalHeader><h1>Clear all cart contents.</h1></ModalHeader>
+//   <ModalCloseButton backgroundColor="#00000000" border="none" _focus={styleButtonFocus}><SvgIcon size="1em" path={times} /></ModalCloseButton>
+//     <ModalBody>
+//       Are you sure you want remove all items ({Object.keys(selectedGranules).length} total)?
+//     </ModalBody>
+//
+//     <ModalFooter>
+//
+//       <Button
+//       style={styleButton}
+//       styleFocus={styleButtonFocus}
+//       onClick={onClose}>Never mind</Button>
+//       <Button
+//       style={styleButton}
+//       styleFocus={styleButtonFocus}
+//       onClick={() => {
+//                 deselectAllGranules()
+//                 onClose()
+//               }}>
+//               Yes
+//       </Button>
+//     </ModalFooter>
+//   </ModalContent>
+// </ModalOverlay>
+// </Modal>
 
   return (
     <div style={styleCenterContent}>
       <Meta title="File Access Cart" robots="noindex" />
+      <Confirmation isOpen={isOpen} onClose={onClose} title="Clear all cart contents." question={`Are you sure you want remove all items (${Object.keys(
+          selectedGranules
+        ).length} total)?`}  yesAction={deselectAllGranules} yesText="Empty cart" noText="Never mind"/>
 
-      <Modal isOpen={isOpen} onClose={onClose} isCentered >
-        <ModalOverlay backgroundColor="#5d5d5d94">
-        <ModalContent maxWidth="28em" borderRadius='0.309em' padding="0.309em">
-          <ModalHeader><h1>Clear all cart contents.</h1></ModalHeader>
-        <ModalCloseButton backgroundColor="#00000000" border="none" _focus={styleButtonFocus}><SvgIcon size="1em" path={times} /></ModalCloseButton>
-          <ModalBody>
-            Are you sure you want remove all items ({Object.keys(selectedGranules).length} total)?
-          </ModalBody>
 
-          <ModalFooter>
-
-            <Button
-            style={styleButton}
-            styleFocus={styleButtonFocus}
-            onClick={onClose}>Never mind</Button>
-            <Button
-            style={styleButton}
-            styleFocus={styleButtonFocus}
-            onClick={() => {
-                      deselectAllGranules()
-                      onClose()
-                    }}>
-                    Yes
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </ModalOverlay>
-      </Modal>
 
       <div style={styleCartListWrapper}>
         {cartActionsWrapper}
