--- client/src/components/common/dialog/OneStopDialog.css
@@ -1,58 +1,0 @@
-.onestop-dialog-base {
-}
-
-.onestop-dialog-overlay {
-}
-
-.onestop-dialog-element {
-  font-size: 1em;
-  position: fixed; /* Stay in place */
-  z-index: 2000; /* Sit on top */
-  left: 0;
-  top: 0;
-  width: 100%; /* Full width */
-  height: 100%; /* Full height */
-  overflow: auto; /* Enable scroll if needed */
-  background-color: rgb(0, 0, 0); /* Fallback color */
-  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
-  padding: 0;
-  margin: 0;
-}
-
-.onestop-dialog-document {
-  font-size: 1em;
-  background-color: #fefefe;
-  margin: 15% auto; /* 15% from the top and centered */
-  padding: 0.618em 1em;
-  border: 1px solid black;
-  border-radius: 0.618em;
-  width: 24em;
-  transition: background-color 2s ease;
-  box-shadow: 2px 2px 4px rgba(50, 50, 50, 0.75);
-}
-
-.onestop-dialog-document-open {
-  background-color: green;
-}
-
-.onestop-dialog-title {
-  font-size: 1em;
-  font-family: 'Merriweather', 'Cambria', 'Times New Roman', 'Times', serif;
-  text-decoration: underline;
-}
-
-.onestop-dialog-closeButton {
-  border: none;
-  background: none;
-  padding: 0;
-  margin: 0.618em;
-  width: 1em;
-  height: 1em;
-  float: right;
-}
-
-.onestop-dialog-closeButton:focus {
-  outline: 2px dashed black;
-  outline-offset: 0.309em;
-  background: none;
-}
